backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
python-jose[cryptography]==3.5.0
pydantic-settings==2.10.1
email-validator==2.2.0
pytest==8.4.1
httpx==0.28.1
sqlalchemy==2.0.43
alembic==1.16.5
psycopg[binary]==3.2.10
